change createdTime to long

+++ serverapitest/src/cn/edu/fudan/blueflamingo/HandinHandServerHelper/Answer.java
@@ -7,11 +7,11 @@ public class Answer {
 	public int score2 = 0;
 	public int uid = 0;
 	public int qid = 0;
-	public int createdTime = 0;
+	public long createdTime = 0;
 	public String picture = "";
 	public int parentAid = 0;
 	
-	public Answer(int id,String content,int score1,int score2,int uid,int qid,int createdTime, String picture,int parentAid){
+	public Answer(int id,String content,int score1,int score2,int uid,int qid,long createdTime, String picture,int parentAid){
 		this.id = id;
 		this.content = content;
 		this.score1 = score1;
