+++ serverapitest/src/cn/edu/fudan/blueflamingo/HandinHandServerHelper/ExAnswer.java
@@ -3,7 +3,7 @@ package cn.edu.fudan.blueflamingo.HandinHandServerHelper;
 public class ExAnswer extends Answer{
 	public String username = "";
 	public String userHead = "";
-	public ExAnswer(int id,String content,int score1,int score2,int uid,int qid,int createdTime, String picture,int parentAid,String username,String userHead){
+	public ExAnswer(int id,String content,int score1,int score2,int uid,int qid,long createdTime, String picture,int parentAid,String username,String userHead){
 		this.id = id;
 		this.content = content;
 		this.score1 = score1;
